fixed an mentor detect issue
